requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-crispy-forms==1.7.2
django-simple-search==1.0.2
Faker==1.0.7
idna==2.8
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.3
